adds creating user in postgres db using prisma
installed prisma to manage connecting to and querying the postgres
database.

from the signup flow, we are now able to check if the email
received is already in use. if it is, we will throw a 409 error,
otherwise we will allow the user to be created (as long as all other
checks pass of course). the user is then created in the db with a
password that is hashed via bcrypt.