updates nav so does not require js to toggle open/close
the nav toggle is now purely html and css so is more accessible and does
not require js. the only js that remians is a small listener that
adds/removes a class from the nav overlay to prevent it from
transitioning while the screen is being resized.

this update also adds a desktop nav instead of reusing the mobile nav.