refactors signout with signout landing page
the user and authenticated stores were not being reset when we were just
redirecting from the server. to solve this, I add "You have been signed
out" page. When the user lands on this page, it confirms with them that
they have indeed logged out, but it also let's us clear the user and
authenticated stores.