fixes validation errors in add recipe form
errors were not being displayed because of logic used to determine if
an error existed, so simplified that logic to ensure error messages are
identified and displayed appropriately.

also updated the add recipe service method with the correct error
message for the recipe description when it was either too long or too
short.